+++ test/credentials.test.js
@@ -89,7 +89,7 @@ describe("credential prompts", () => {
 
   it("settles the outer prompt promise on secret prompt errors", () => {
     const source = fs.readFileSync(
-      path.join(import.meta.dirname, "..", "bin", "lib", "credentials.js"),
+      path.join(import.meta.dirname, "..", "src", "lib", "credentials.ts"),
       "utf-8",
     );
 
@@ -100,7 +100,7 @@ describe("credential prompts", () => {
 
   it("re-raises SIGINT from standard readline prompts instead of treating it like an empty answer", () => {
     const source = fs.readFileSync(
-      path.join(import.meta.dirname, "..", "bin", "lib", "credentials.js"),
+      path.join(import.meta.dirname, "..", "src", "lib", "credentials.ts"),
       "utf-8",
     );
 
@@ -114,7 +114,7 @@ describe("credential prompts", () => {
     expect(credentials.normalizeCredentialValue("  nvapi-good-key\r\n")).toBe("nvapi-good-key");
 
     const source = fs.readFileSync(
-      path.join(import.meta.dirname, "..", "bin", "lib", "credentials.js"),
+      path.join(import.meta.dirname, "..", "src", "lib", "credentials.ts"),
       "utf-8",
     );
     expect(source).toMatch(/while \(true\) \{/);
@@ -124,7 +124,7 @@ describe("credential prompts", () => {
 
   it("masks secret input with asterisks while preserving the underlying value", () => {
     const source = fs.readFileSync(
-      path.join(import.meta.dirname, "..", "bin", "lib", "credentials.js"),
+      path.join(import.meta.dirname, "..", "src", "lib", "credentials.ts"),
       "utf-8",
     );
